requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.0
pydantic>=2.10.6
sqlalchemy>=2.0.27
alembic>=1.16.5
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
supabase>=2.3.0